voice-assistant-backend/requirements.txt
aenum==3.1.15
aiofiles==23.2.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2
attrs==24.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
dataclasses==0.6
dataclasses-json==0.6.7
deepgram-sdk==3.7.3
deprecation==2.1.0
distro==1.9.0
dnspython==2.7.0
elevenlabs==1.9.0
fastapi==0.115.3
frozenlist==1.4.1
greenlet==3.1.1
groq==0.11.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
hypothesis==6.115.6
idna==3.10
iniconfig==2.0.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.3
langchain-core==0.3.10
langchain-groq==0.2.0
langchain-text-splitters==0.3.0
langsmith==0.1.134
marshmallow==3.22.0
motor==3.6.0
multidict==6.1.0
mypy-extensions==1.0.0
numpy==1.26.4
orjson==3.10.7
packaging==24.1
pluggy==1.5.0
propcache==0.2.0
PyAudio==0.2.14
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
pymongo==4.9.2
pytest==7.2.2
pytest-asyncio==0.20.3
pytest-mock==3.10.0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.1
SQLAlchemy==2.0.35
starlette==0.41.0
tenacity==8.5.0
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
websockets==12.0
yarl==1.15.2
schedule